fé

+++ services/agendaService/src/modules/Agenda/Aplication/UseCases/commands/room/deleteRoom.py
@@ -1,21 +1,25 @@
 from src.modules.agenda.aplication.ports.events.BusPort import BusPort
+from src.modules.agenda.aplication.events.RoomEvent import DeleteRoomEvent
 from src.modules.agenda.domain.entities.Room import Room
 from src.modules.agenda.aplication.ports.repository.RoomRepositoryPort import RoomRepositoryPort
 
-class DeleteRoom:
+class DeleteRoomUseCase:
     def __init__(self, repository: RoomRepositoryPort, bus: BusPort):
         self._repository = repository
         self._bus= bus
-    def execute(self, room_id: int) -> bool:
+    async def execute(self, room_id: str) -> bool:
         try:
-            room =self._repository.getRoom(room_id) 
+            room = await self._repository.getRoom(room_id)
+            if not isinstance(room, Room):
+                return False
             roomEntity = Room(name = room.name, disponibility = room.disponibility, rules = room.rules)
             
             if(room and roomEntity.delete()):
-                self._repository.deleteRoom(room_id)
-                self._bus.emit('roomDeleted', room_id)
+                await self._repository.deleteRoom(room_id)
+                self._bus.emit(DeleteRoomEvent(str(room_id)))
                 return True
         
         except Exception as e:
-            return 
+            return False
+        return False
